split out humungous file
