+++ packages/server/src/sdk/workspace/ai/agentLogs/index.ts
@@ -1,0 +1,18 @@
+export type {
+  CreateSessionLogIndexerInput,
+  IndexAgentLogOperationInput,
+  SessionLogIndexer,
+} from "@budibase/types"
+
+export {
+  clearOldHistory,
+  getExpiredSessions,
+  oldestLogDate,
+} from "./sessionIndex"
+export {
+  addSessionLog,
+  createSessionLogIndexer,
+  initLogIndexQueue,
+} from "./indexQueue"
+export { fetchSessions } from "./sessions"
+export { fetchRequestDetail, fetchSessionDetail } from "./detail"
